// template/base/src/stores/app/action-types.ts
export enum AppActionTypes {
  TOGGLE_SIDEBAR = "TOGGLE_SIDEBAR",
  CLOSE_SIDEBAR = "CLOSE_SIDEBAR",
  TOGGLE_DEVICE = "TOGGLE_DEVICE",
  SET_LANGUAGE = "SET_LANGUAGE",
  SET_SIZE = "SET_SIZE",
  GET_DOORINFO = "GET_DOORINFO",
  GET_APPINFO = "GET_APPINFO",
  GET_USERINFO = "GET_USERINFO",
  GET_ROUTEFROMAPI = "GET_ROUTEFROMAPI",
  SET_FINALLYROUTE = "SET_FINALLYROUTE",
  LOGIN_OUT = "LOGIN_OUT",
}